requirement.txt
fastapi
uvicorn[standard]
python-multipart
langchain-groq
langchain
langchain-community
pydantic
python-dotenv
pymupdf
beautifulsoup4
streamlit
requests
pandas
openpyxl